Rework how session hash keys are created
This changes the hash key to created and stored when the key object
is created. This unfortunately also changes the key object a bit which
results in the hash being different, and some of the tests to change
due to ordering.

The change also resolves a fairly significant performance issue where
hashing too much was slowing down session insertion.